vaapipostproc: consider direction in src caps for size flag
GST_VAAPI_POSTPROC_FLAG_SIZE should take into consideration
when WxH is swapped (i.e. HxW) by video-direction.  That is,
it should not be set when direction results in 90 or 270
degree rotation.  This fixes an issue in src_event where
90r, 90l, ur-ll and ul-lr cases could not be reached due
to the first condition of GST_VAAPI_POSTPROC_FLAG_SIZE is
set.