libs: decoder: do not directly return error in post unit.
Some stream has error unit after the AU. The decoder can successfully
decode the picture and insert it into DPB. But the post unit error
will make that frame dropped. The later output of the picture will
ref that frame and crash.
Still continue to handle the frame and just report a post unit error
can solve this.