libs: replace gstvaapiobject using GstMiniObject.
The GstVaapiMiniObject is obsolete and need to use GstMiniObject to
replace it. The gstvaapiobject is the base class of many useful vaapi
objects, such as image, surface and texture. we still define a common
class as the first part of all these objects and use GstMiniObject to
replace all GstVaapiMiniObject inside it, which can make all the
objects follow standard gst framework and make the code clean.