plugins: encode: Motify encode's sink caps.
The old static way of template caps is not very correct. It loses
a lot of supported formats and reports a lot of totally unsupported
formats in the caps. We need to use the refined video format string
when video format map is created. This makes the template caps more
precise.