plugin: encode: modify get_profiles to list all possible profiles.
The current get_profile just gives one possible profile to detect the
input surface of the encode, which may not be enough if we want to
support more profiles. For example, if we want to support HEVC 4:4:4
profile, the input of encode should be VYUA rather than NV12 in HEVC
main profile. The command line:
  gst-launch-1.0 videotestsrc num-buffers=200 ! capsfilter \
  caps=video/x-raw,format=VUYA,width=800,height=600 ! vaapih265enc \
  tune=low-power init-qp=30 ! fakesink
can not work because vaapih265enc just report NV12 caps, we need to
specify the profile obviously like:
  gst-launch-1.0 videotestsrc num-buffers=200 ! capsfilter \
  caps=video/x-raw,format=VUYA,width=800,height=600 ! vaapih265enc \
  tune=low-power init-qp=30 ! capsfilter caps=video/x-h265, \
  profile=main-444 ! fakesink

The encode should have the ability to choose the profile based on input
format. If the input video format is VUYA, the main-444 profile should
be auto selected.

We modify get_profile to get_profiles to make it return all possible
supported profiles. All these profiles will be used to create test
context and merge the supported formats together. So for this case,
both NV12 and VUYA should be returned.