libs: encoder: get surfaces resolution the same time with formats.
We can get all the information about the video format at one shot
when we create the test context for getting the supported formats.
The current manner to get the width and height is very verbose, it
calls the function gst_vaapi_profile_caps_append_encoder to create
another temp context to detect the resolution information.