plugin: encode: List all possible profiles to detect input formats.
The current get_profile just return one possible profile for the encode,
which is not enough.  For example, if we want to support HEVC 4:4:4
profile, the input of encode should be VYUA rather than NV12 in HEVC
main profile. So the command line:

  gst-launch-1.0 videotestsrc num-buffers=200 ! capsfilter \
  caps=video/x-raw,format=VUYA,width=800,height=600 ! vaapih265enc \
  tune=low-power init-qp=30 ! fakesink

can not work because vaapih265enc just report NV12 caps, we need to
specify the profile obviously like:

  gst-launch-1.0 videotestsrc num-buffers=200 ! capsfilter \
  caps=video/x-raw,format=VUYA,width=800,height=600 ! vaapih265enc \
  tune=low-power init-qp=30 ! capsfilter caps=video/x-h265, \
  profile=main-444 ! fakesink

The encode should have the ability to choose the profile based on input
format automatically. If the input video format is VUYA, the main-444
profile should be auto selected.
We now just let get_profile_allowed return GST_VAAPI_PROFILE_UNKNOWN if
no profile specified downstream or can not parse it. This will let the
function gst_vaapi_encoder_get_surface_attributes collects the surface's
attributes for all profiles which belong to the current encoder's codec.
So for this case, both NV12 and VUYA should be returned.