plugin: encode: List all possible profiles to detect input formats.
The current get_profile just return one possible profile for the encode,
which is not enough.  For example, if we want to support HEVC 4:4:4
profile, the input of encode should be VYUA rather than NV12 in HEVC
main profile. So the command line:

  gst-launch-1.0 videotestsrc num-buffers=200 ! capsfilter \
  caps=video/x-raw,format=VUYA,width=800,height=600 ! vaapih265enc \
  tune=low-power init-qp=30 ! fakesink

can not work because vaapih265enc just report NV12 in sink caps, we need
to specify the profile obviously like:

  gst-launch-1.0 videotestsrc num-buffers=200 ! capsfilter \
  caps=video/x-raw,format=VUYA,width=800,height=600 ! vaapih265enc \
  tune=low-power init-qp=30 ! capsfilter caps=video/x-h265, \
  profile=main-444 ! fakesink

The encode should have the ability to choose the profile based on input
format automatically. If the input video format is VUYA, the main-444
profile should be auto choosed.

We modify to let get_allowed_profiles return an array of all supported
profiles based on downstream's allowed caps, or return NULL if no profile
specified downstream or the encode do not provide get_profile() function.
The function gst_vaapi_encoder_get_surface_attributes collects the surface's
attributes for profiles allowed, or for all profiles which belong to the
current encoder's codec.

So for this case, both NV12 and VUYA should be returned.

TODO: some codec like VP9, need to implement the get_profile() function.