plugin: encode: no need to wait if surfaceless.
The notification in gst_vaapidecode_release have some overhead, we
do not really need it if the decoder context is surfaceless. And
also, for surfaceless context, if no more surface is available, it
means a serious problem such as GPU memory leak. In order not to
cause bigger problem(such as system hang for some unrobust driver),
we just fail the pipeline here.