vaapipostproc: use sink resolution to calculate src colorimetry
The default output colorimetry was decided by the output
resolution, which is too naive when doing VPP cropping
and/or scaling.

Thus, if downstream (src) does not specify a colorimetry
then we use the sink resolution to create the default
colorimetry instead of the output resolution.  This ensures
that VPP uses compatible input and output colorimetry
regardless of the output resolution when colorimetry is
unspecified by downstream.