libs: encoder: WA: set HRD params before BRC params
This is a work around for intel-media-driver bug
https://github.com/intel/media-driver/issues/865

This driver will force the BRC to CBR for HEVCe when it
parses the HRD params.  Thus, any BRC method submitted
in the BRC params prior to parsing the HRD params will
be lost.  Therefore, VBR can't be effectively enabled
if BRC params precede the HRD params.

To work around this issue, set the HRD params before
the BRC params so the driver will parse BRC "after"
HRD params.