libs: decoder: create a test context from caps
Context is created deep down in the internal decoders just before
pushing the first buffer, since streams are parsed internally. This
design doesn't help to enhance caps negotiation, since it's required
to know, as soon as possible, if the incoming stream, previously
parsed by a parser, can be processed.

The solution presented in this patch is to create a temporal VA
context from the sink caps, if the internal decoder hasn't yet created
its own.

The approach has the problem if the context created from caps is
different from the one created by the internal decoder, leading to the
generation of a wrong set of allowed src caps. We seems this danger
has small change to appear.