vaapidecode: split update_src_caps() and set_output_state()
Since we need to configure the output surfaces from the negotiated
caps features and color format, it is required the split
update_src_caps() in two functions: one for negotiating the src pad
caps and other to actually set the output state.

The first one will keep the same function name, and will store an
initial srcpad caps, not exact, but with the negotiated caps feature
and color format. Later on, at set_output_state() the srpad caps
will be updated with its final configuration.