libs: video-format: remove encoded from vaapi formats
the purpose of ENCODED format was an initial idea for crypted media,
but I'm not aware of nobody using it and I doubt it actually works
with the current VA-API encrypted handling.

In order to simplify further code this patch just get rid of ENCODED
from the static caps of gstreamer-vaapi, thus it won't be negotiated.