vaapidecode, vaapipostproc: simplify downstream negotiation
The function gst_vaapi_find_preferred_caps_feature() was too complex
for its purpose: since gst_caps_fixate() could choose a not-optimal
capsfeature, the function returned the "best" capsfeature and the
first color format.

The function, originally, did a peer caps query and an caps intersect,
but all that is done already by gst_pad_get_allowed_caps(), and later
choose the first color format in way that can be simplified using
gst_caps_fixate() in the proper caps.

The problem was that both vaapidecode and vaapipostproc uses the
function, and gst_pad_get_allowed_caps() is problematic for
vaapipostproc because the caller function is locked.

Thus, first of all, the patch moves out the
gst_pad_get_allowed_caps(), then
gst_vaapi_find_preferred_caps_feature() only find, from a GstCaps set,
the "best" caps feature, and once found, it fixates it to find the
first color format.

In the case of vaapipostproc, before calling
gst_pad_get_allowed_caps() the lock is unlocked and later locked again
to avoid a deadlock. And did an intersection with the othercaps from
fixate_caps() vmethod.

In the case of vaapidecode, also call gst_pad_get_allowed_caps(), and
later an intersection with a reasonable assumption of the caps that
are going to be generated by the decoder afterwards, so the decoder
could known, as soon as possible, the caps feature and the color
format.