vaapipluginutil: simplify gst_vaapi_find_preferred_caps_feature()
Rather than a O^2 search for a capfeature in an ordered list of features, now
it only searchs for memory:VASurface, because it is the preferred. Otherwise
the caps are fixed and, if the capsfeature VASurface isn't found, a simple
search for the other preferred caps are search in the fixated caps.

This is better than before because it is simpler, but also is more aligned
with the caps negotiation design in GStreamer.