libs: decoder: naive approach to order src formats
Output preference is dependent of input chroma depth. This is
supposed to be handled by the driver when querying the surfaces
formats given a context, but it is not assured.

This patch sorts the reported array of formats from the drivers
given an input chroma depth. The comparation function implements a
naive approach by sign switching.