plugin: allocator: no image pool needed for GstVaapiVideoAllocator.
We do not need to allocate a image pool inside the GstVaapiVideoAllocator
to cache the image. In fact, the image can be hold in GstVaapiVideoMemory,
and that memory and the buffer will be cached by the gstbufferpool.
So we change the policy to:
1. When not "use_native_formats", that is, the image is generated by the
   vaDeriveImage(), we never cache it. We generate it when map() and destroy
   it when unmap().
2. If "use_native_formats", the image is generated by the vaCreateImage() at
   the first time of map(). We will store it in the GstVaapiVideoMemory until
   the memory is destroyed.
We also change the unmap() to call vaPutImage() if needed. So when a image is
mapped and changed, we will sync it with the surface when unmap().