plugin: allocator: delete surface_pool for GstVaapiVideoAllocator.
There is no need to use the surface_pool to cache the surfaces used
by the GstVaapiVideoMemory. The gstbufferpool can cache the gstbuffer
and also the surfaces stored in gstbuffer.
We now devide the GstVaapiVideoMemory into two subtypes.
1. It creates and holds the surface itself. We just allocate the surface
when creating the GstVaapiVideoMemory and never reset the surface. The
surface will be destroyed when GstVaapiVideoMemory is freed.
2. It uses the surface specified by the user. This mode is useful such
as the decoder wants to create a gstbuffer with an already decoded picture.
We just ref the surface when creating and unref the surface when resetting.
User can use gst_vaapi_video_buffer_set_video_mem_surface_proxy() to change
the underlying surface when needed.